// common-ms/flashex-webapp/flashex-webapp/src/app/trip-management/interfaces/triplog.ts
export interface ITripLog {
  tripItineraryId: string;
  tripStart: Date;
  tripEnd: Date;
  packetLogs: {
    packetId: string;
    deliveryAddress: {
      addressLine1: string;
      city: string;
      state: string;
      latitude: number,
      longitude: number,
      pincode: number
    };
    deliveryDescription: string;
    packetStatus: string;
  };
  plannedStartTime: Date;
  plannedEndTime: Date;
  originAddress: string;
}